Moved factory and algorithms into new meta type
