+++ src/algorithms/meta/CollectionCollector.h
